Centralize logic of postprocess memory to util
